Passed testing and followed structure of tests
